PRSD-660: Renames searchModels folder to filterPanelModels
