+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/filterPanelModels/FilterPanelViewModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.viewModels.searchModels
+package uk.gov.communities.prsdb.webapp.models.viewModels.filterPanelModels
 
 import jakarta.servlet.http.HttpServletRequest
 import uk.gov.communities.prsdb.webapp.helpers.URIQueryBuilder
